refactor: massive refactoring around the events feature
